REname CPCkeymap.h to libretro-keymap.h
